Move Dev Handler to specific file
